requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Document processing
PyMuPDF==1.23.9
pdfplumber==0.10.3
python-multipart==0.0.6
python-docx==1.1.0

# Machine Learning and Embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2

# Local LLM Integration - Ollama
ollama>=0.1.0

# HTTP and requests
requests==2.31.0
httpx==0.25.2

# Note: hashlib, tempfile, logging, concurrent.futures are built-in Python modules